+++ CHANGES.rst
@@ -37,6 +37,7 @@ Changelog
     `MeetingConfig.useGroupsAsCategories` was `True`, now it returns only categories, moreover parameter `classifiers` is
     renamed to `catType` that may be `all`/`categories`/`classifiers`.
 - In every migrations, call `cleanRegistries` at the end by default so `JS/CSS` are recompiled
+- Make sure `++resource++plone.app.jquerytools.dateinput.js` is enabled in `portal_javascripts`
 - Improved `BaseDGHV.print_signatories_by_position` to add more use cases
 - Added tests for `BaseDGHV.print_signatories_by_position`